integrated checkGroup
