--- src/components/group/index.js
@@ -1,1 +1,0 @@
-export * from './checkboxGroup'
